fix: 동시성 제어를 위해 브로드캐스트 로직 분리, 테스트코드 수정 전
