+++ src/main/java/com/back/global/websocket/service/WebSocketSessionManager.java
@@ -1,30 +1,30 @@
 package com.back.global.websocket.service;
 
-import com.back.domain.studyroom.dto.RoomBroadcastMessage;
 import com.back.global.exception.CustomException;
 import com.back.global.exception.ErrorCode;
 import com.back.global.websocket.dto.WebSocketSessionInfo;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.context.annotation.Lazy;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 
 import java.time.Duration;
 import java.util.Set;
 import java.util.stream.Collectors;
 
+/**
+ * WebSocket 세션 관리 전용 서비스
+ * - Redis 기반 세션 상태 관리
+ * - 방 입장/퇴장 처리
+ * - 온라인 사용자 조회
+ * - 브로드캐스트는 WebSocketBroadcastService로 분리
+ */
 @Slf4j
 @Service
 @RequiredArgsConstructor
 public class WebSocketSessionManager {
 
     private final RedisTemplate<String, Object> redisTemplate;
-
-    // 브로드캐스트 부분에서 순환 참조가 생겨서 Lazy 주입
-    @Lazy
-    private final SimpMessagingTemplate messagingTemplate;
 
     // Redis Key 패턴
     private static final String USER_SESSION_KEY = "ws:user:{}";
@@ -248,86 +248,6 @@ public class WebSocketSessionManager {
             String userKey = USER_SESSION_KEY.replace("{}", userId.toString());
             redisTemplate.delete(userKey);
             redisTemplate.delete(sessionKey);
-        }
-    }
-
-    // ======================== 브로드캐스트 기능 ========================
-
-    /**
-     * 특정 방의 모든 온라인 사용자에게 메시지 브로드캐스트
-     */
-    public void broadcastToRoom(Long roomId, RoomBroadcastMessage message) {
-        try {
-            Set<Long> onlineUsers = getOnlineUsersInRoom(roomId);
-
-            if (onlineUsers.isEmpty()) {
-                log.debug("브로드캐스트 대상이 없음 - 방: {}", roomId);
-                return;
-            }
-
-            // 방 전체 토픽으로 브로드캐스트
-            String destination = "/topic/rooms/" + roomId + "/updates";
-            messagingTemplate.convertAndSend(destination, message);
-
-            log.info("방 브로드캐스트 완료 - 방: {}, 타입: {}, 대상: {}명",
-                    roomId, message.getType(), onlineUsers.size());
-
-        } catch (Exception e) {
-            log.error("방 브로드캐스트 실패 - 방: {}, 타입: {}", roomId, message.getType(), e);
-            throw new CustomException(ErrorCode.WS_REDIS_ERROR);
-        }
-    }
-
-    /**
-     * 특정 사용자에게 개인 메시지 전송
-     */
-    public void sendToUser(Long userId, String destination, Object message) {
-        try {
-            if (isUserConnected(userId)) {
-                messagingTemplate.convertAndSendToUser(
-                        userId.toString(),
-                        destination,
-                        message
-                );
-
-                log.debug("개인 메시지 전송 완료 - 사용자: {}, 목적지: {}", userId, destination);
-            } else {
-                log.debug("오프라인 사용자에게 메시지 전송 시도 - 사용자: {}", userId);
-            }
-        } catch (Exception e) {
-            log.error("개인 메시지 전송 실패 - 사용자: {}", userId, e);
-        }
-    }
-
-    /**
-     * 방의 온라인 멤버 목록 업데이트 브로드캐스트
-     */
-    public void broadcastOnlineMembersUpdate(Long roomId) {
-        try {
-            Set<Long> onlineUsers = getOnlineUsersInRoom(roomId);
-
-            // 온라인 사용자 ID 목록을 브로드캐스트
-            RoomBroadcastMessage message = RoomBroadcastMessage.onlineMembersUpdated(
-                    roomId,
-                    onlineUsers.stream().toList()
-            );
-
-            broadcastToRoom(roomId, message);
-
-        } catch (Exception e) {
-            log.error("온라인 멤버 목록 브로드캐스트 실패 - 방: {}", roomId, e);
-        }
-    }
-
-    /**
-     * 방 상태 변경 알림 브로드캐스트
-     */
-    public void broadcastRoomUpdate(Long roomId, String updateMessage) {
-        try {
-            RoomBroadcastMessage message = RoomBroadcastMessage.roomUpdated(roomId, updateMessage);
-            broadcastToRoom(roomId, message);
-        } catch (Exception e) {
-            log.error("방 상태 변경 브로드캐스트 실패 - 방: {}", roomId, e);
         }
     }
 }
